dashboard/pyproject.toml
[project]
name = "dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.41",
    "taipy>=3.0.0",
]
